# src/IO.Swagger/Models/Collections.cs
/*
 * Index Card API
 *
 * This is a simple server for loading and saving index cards
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{
    /// <summary>
    /// Model for a collection of index-cards
    /// </summary>
    [DataContract]
    public partial class Collections : IEquatable<Collections>
    { 
        /// <summary>
        /// Gets or Sets CollectionId
        /// </summary>

        [DataMember(Name="collectionId")]
        public Guid? CollectionId { get; set; }

        /// <summary>
        /// Gets or Sets UserId
        /// </summary>

        [DataMember(Name="userId")]
        public Guid? UserId { get; set; }

        /// <summary>
        /// Gets or Sets CollectionName
        /// </summary>

        [DataMember(Name="collectionName")]
        public string CollectionName { get; set; }

        /// <summary>
        /// Gets or Sets NumberOfCards
        /// </summary>

        [DataMember(Name="numberOfCards")]
        public int? NumberOfCards { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Collections {\n");
            sb.Append("  CollectionId: ").Append(CollectionId).Append("\n");
            sb.Append("  UserId: ").Append(UserId).Append("\n");
            sb.Append("  CollectionName: ").Append(CollectionName).Append("\n");
            sb.Append("  NumberOfCards: ").Append(NumberOfCards).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Collections)obj);
        }

        /// <summary>
        /// Returns true if Collections instances are equal
        /// </summary>
        /// <param name="other">Instance of Collections to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Collections other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    CollectionId == other.CollectionId ||
                    CollectionId != null &&
                    CollectionId.Equals(other.CollectionId)
                ) && 
                (
                    UserId == other.UserId ||
                    UserId != null &&
                    UserId.Equals(other.UserId)
                ) && 
                (
                    CollectionName == other.CollectionName ||
                    CollectionName != null &&
                    CollectionName.Equals(other.CollectionName)
                ) && 
                (
                    NumberOfCards == other.NumberOfCards ||
                    NumberOfCards != null &&
                    NumberOfCards.Equals(other.NumberOfCards)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (CollectionId != null)
                    hashCode = hashCode * 59 + CollectionId.GetHashCode();
                    if (UserId != null)
                    hashCode = hashCode * 59 + UserId.GetHashCode();
                    if (CollectionName != null)
                    hashCode = hashCode * 59 + CollectionName.GetHashCode();
                    if (NumberOfCards != null)
                    hashCode = hashCode * 59 + NumberOfCards.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Collections left, Collections right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Collections left, Collections right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}