# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Index Card API",
    "description" : "This is a simple server for loading and saving index cards",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://dev.index-cards.com/v1",
    "description" : "Dev Server"
  }, {
    "url" : "https://prod.index-cards.com/v1",
    "description" : "Prod Server"
  } ],
...
  "paths" : {
...
    "/collections/{userId}" : {
      "get" : {
        "tags" : [ "collections" ],
        "summary" : "Get all collections of user",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully pulled all collections",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Collections"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "500" : {
            "description" : "internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "collections" ],
        "summary" : "Create a new collection",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "requestBody" : {
          "description" : "the new collection",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Collections"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successfully created new collection",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/collections/{collectionId}" : {
      "delete" : {
        "tags" : [ "collections" ],
        "summary" : "delete the selected collection",
        "parameters" : [ {
          "name" : "collectionId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "successfully deleted collection"
          },
          "404" : {
            "description" : "collection not found"
          },
          "500" : {
            "description" : "internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Collections" : {
        "type" : "object",
        "properties" : {
          "collectionId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          },
          "collectionName" : {
            "type" : "string",
            "example" : "Geography"
          },
          "numberOfCards" : {
            "type" : "integer",
            "example" : 20
          }
        },
        "description" : "Model for a collection of index-cards",
        "example" : {
          "numberOfCards" : 20,
          "collectionId" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243",
          "userId" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243",
          "collectionName" : "Geography"
        }
      },
...
      "inline_response_200" : {
        "type" : "object",
        "properties" : {
          "collectionName" : {
            "type" : "string",
            "example" : "Geography"
          }
        },
        "example" : {
          "collectionName" : "Geography"
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}